refactor: 애플리케이션 패키지 생성
