+++ src/main/java/racingcar/application/RacingCarApiApplication.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
