+++ src/main/java/racingcar/application/RacingCarConsoleApplication.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.application;
 
 import java.util.Scanner;
 import racingcar.controller.RacingGameController;
